feat: EAD Padaria — OAuth Google, mídias, minigames, PWA, documentação e melhorias de UX
Autenticação e API
- Login em etapas (check-email), registro e JWT com refresh estendido
- OAuth Google (Socialite): rotas web, callback e redirecionamento ao frontend
- AuthController e rotas em routes/api.php; migration google_id em users
- CORS, APP_URL/APP_FRONTEND_URL e configs alinhadas a produção

Mídias e conteúdo
- Media, MediaCategory, controllers, policies e rotas REST
- Frontend: MediaManager, MediaPickerDialog, UrlFieldWithPicker, media.service
- Integração com formulários e WysiwygEditor

Minigames
- GameAdminController CRUD (gerente); novos tipos de jogo e telas (player, manager)
- game.admin.service, MinigamePlayer, componentes de jogos

Editor e UI
- Bloco HTML customizado no Quill (iframe srcdoc) para script/style
- HtmlInsertDialog; estilos de toolbar; PageLoader/AppLayout/Suspense
- Login: animação entre etapas, Google button, correções de overflow/outline
- Sidebar com modal (scroll lock): ajustes em index.css
- AlertPopupContext, OfflineIndicator, PWAUpdatePrompt, ícones PWA

Documentação
- docs/: visão geral, requisitos, arquitetura, modelo de dados, API, frontend
- COMPILADO.md e README raiz com deploy e links
- CHANGELOG.md; GOOGLE_OAUTH_SETUP e demais guias em docs/

Outros
- public/.htaccess (ModSecurity/OAuth); DevEnvironmentServiceProvider
- Build frontend (dist), service worker, dependências composer/npm
- Scribe atualizado; remoção push.debug.ts

Made-with: Cursor

+++ app/Services/BadgeAwardService.php
@@ -115,7 +115,7 @@ class BadgeAwardService
 
         foreach ($badges as $badge) {
             $params = $badge->criteria_params ?? [];
-            $minScore = $params['min_score'] ?? 0;
+            $minScore = (int) ($params['min_score'] ?? 0);
             if ($score >= $minScore && !$user->badges()->where('badge_id', $badge->id)->exists()) {
                 $user->badges()->attach($badge->id, [
                     'course_id' => $game->course_id,
